Refactor payment handling: consolidate PaymentInput and update related classes
