+++ source/api/BogusData/BogusUpdateRegistrationInput.cs
@@ -14,6 +14,6 @@ internal sealed class BogusUpdateRegistrationInput
         RuleFor(input => input.SweeperIds, f => f.Make(f.Random.Number(0, 3), () => SquadId.New()).OrNull(f));
         RuleFor(input => input.SuperSweeper, f => f.Random.Bool().OrNull(f));
 
-        RuleFor(input => input.PaymentConfirmation, f => f.Finance.LitecoinAddress().OrNull(f));
+        RuleFor(input => input.Payment, _ => new BogusPaymentInput().Generate());
     }
 }
